WUI: setAffiliation support for combination of GROUP and VO manager
The old version did not support if a user had both roles of VO_MANAGER and GROUP_MANAGER at the same time. Now the user can choose from VOs and GROUPS that one is manager of from one select. Also if the user has only one role the select stays the same.