Add forceAuthn for identity consolidation workflow
When the identity consolidation workflow starts it necessary to send
forceAuth parameter to SP which will pass it to the proxy IdP.

If the forceAuth flag is not present, the SP have to be configured to
send logout to proxy and the proxy have to do the logout as well. Not
all proxies are supporting this, so for that using forceAuthn is
required.