BugFix - username filed was always disabled
* There was a problem with the `enabled` property of the `Username`
object. The value of this property was set to true, when instance of
this class was created. This property was used in the `initWidget` method.
The `initWidget` method was called from a constructor of a super class
and that was a problem. When the super constructor was called, variables
declared in the `Username` class were not initalized. Therefore, the
`enabled` property had the default value - false.
* This caused the username to be always disabled.
* I discovered that the property was not actually needed, co the fix is
just to remove it.